# Don't use Hint for known primitives