Before patch:

```
GHC: Parsing and optimising modules took: 0.550s
GHC: Loading external modules from interface files took: 0.016s
GHC: Parsing annotations took: 0.000s
Hint: interpreting "Clash.Primitives.GHC.Int.intTF"
Hint: interpreting "Clash.Primitives.GHC.Word.wordTF"
Hint: interpreting "Clash.Primitives.GHC.Int.intTF"
Hint: interpreting "Clash.Primitives.Sized.ToInteger.indexToIntegerVerilog"
Hint: interpreting "Clash.Primitives.GHC.Word.wordTF"
Hint: interpreting "Clash.Primitives.GHC.Int.intTF"
Hint: interpreting "Clash.Primitives.GHC.Word.wordTF"
Hint: interpreting "Clash.Primitives.Sized.ToInteger.unsignedToIntegerVerilog"
Hint: interpreting "Clash.Primitives.Sized.ToInteger.signedToIntegerVerilog"
Hint: interpreting "Clash.Primitives.GHC.Int.intTF"
Hint: interpreting "Clash.Primitives.Sized.ToInteger.bvToIntegerVerilog"
Hint: interpreting "Clash.Primitives.GHC.Word.wordTF"
Hint: interpreting "Clash.Primitives.Sized.Vector.indexIntVerilog"
Clash: Parsing and compiling primitives took 5.710s
GHC+Clash: Loading modules cumulatively took 6.373s
Clash: Compiling Trace.topEntity
Clash: Normalisation took 0.000s
Clash: Netlist generation took 0.000s
Clash: Total compilation took 6.382s
```

After:

```
GHC: Parsing and optimising modules took: 0.534s
GHC: Loading external modules from interface files took: 0.016s
GHC: Parsing annotations took: 0.000s
Clash: Parsing and compiling primitives took 1.490s
GHC+Clash: Loading modules cumulatively took 2.136s
Clash: Compiling Trace.topEntity
Clash: Normalisation took 0.007s
Clash: Netlist generation took 0.000s
Clash: Total compilation took 2.145s
```

I.e.:

```
Clash: Parsing and compiling primitives took 5.710s
```

vs

```
Clash: Parsing and compiling primitives took 1.490s
```